Refactor make_json_or_python_to_test_case
