+++ openfisca_core/scenarios.py
@@ -8,7 +8,7 @@ import itertools
 
 import numpy as np
 
-from . import conv, periods, simulations
+from . import conv, periods, simulations, json_to_test
 
 
 def N_(message):
@@ -352,8 +352,7 @@ class AbstractScenario(object):
                 value = value, state = state or conv.default_state)
         return json_to_instance
 
-    def new_simulation(self, debug = False, debug_all = False, reference = False, trace = False,
-            use_set_input_hooks = True, opt_out_cache = False):
+    def new_simulation(self, debug = False, debug_all = False, reference = False, trace = False, use_set_input_hooks = True, opt_out_cache = False):
         assert isinstance(reference, (bool, int)), \
             'Parameter reference must be a boolean. When True, the reference tax-benefit system is used.'
         tax_benefit_system = self.tax_benefit_system
@@ -375,80 +374,13 @@ class AbstractScenario(object):
         return simulation
 
     def make_json_or_python_to_test_case(self, period, repair = False):
-
-        def build_role_parser(role):
-                if role.max == 1:
-                    return conv.test_isinstance((basestring, int))
-                else:
-                    return conv.pipe(
-                        conv.make_item_to_singleton(),
-                        conv.test_isinstance(list),
-                        conv.uniform_sequence(
-                            conv.test_isinstance((basestring, int)),
-                            drop_none_items = True,
-                            ),
-                        conv.default([]),
-                        )
-
-
-        def get_role_parsing_dict(entity):
-            if entity.is_person:
-                return {}
-            else:
-                return {
-                    role.plural: build_role_parser(role)
-                    for role in entity.roles
-                }
-
-
-        def get_entity_parsing_dict(tax_benefit_system):
-            column_by_name = tax_benefit_system.column_by_name
-            return {
-                entity.plural : conv.pipe(
-                            conv.make_item_to_singleton(),
-                            conv.test_isinstance(list),
-                            conv.uniform_sequence(
-                                conv.test_isinstance(dict),
-                                drop_none_items = True,
-                                ),
-                            conv.function(set_entities_json_id),
-                            conv.uniform_sequence(
-                                conv.struct(
-                                    dict(itertools.chain(
-                                        dict(
-                                            id = conv.pipe(
-                                                conv.test_isinstance((basestring, int)),
-                                                conv.not_none,
-                                                ),
-                                            ).iteritems(),
-                                        get_role_parsing_dict(entity).iteritems(),
-                                        (
-                                            (column.name, column.json_to_python)
-                                            for column in column_by_name.itervalues()
-                                            if column.entity == entity
-                                            ),
-                                        )),
-                                    drop_none_values = True,
-                                    ),
-                                drop_none_items = True,
-                                ),
-                            conv.default([]),
-                            )
-                for entity in tax_benefit_system.entities
-            }
-
-
         def json_or_python_to_test_case(value, state = None):
             if value is None:
                 return value, None
             if state is None:
                 state = conv.default_state
 
-            test_case, error = conv.pipe(
-                conv.test_isinstance(dict),
-                conv.struct(get_entity_parsing_dict(self.tax_benefit_system)),
-                )(value, state = state)
-
+            test_case, error = json_to_test.parse_entities(value, self.tax_benefit_system, state)
             return test_case, error
 
         return json_or_python_to_test_case
@@ -748,13 +680,6 @@ def make_json_or_python_to_test(tax_benefit_system):
             }, None
 
     return json_or_python_to_test
-
-
-def set_entities_json_id(entities_json):
-    for index, entity_json in enumerate(entities_json):
-        if 'id' not in entity_json:
-            entity_json['id'] = index
-    return entities_json
 
 
 def iter_over_entity_members(entity_description, scenario_entity):
